Change cache expire to work on every Xth minute.
This is instead of just expiring X minutes from the time the cache was
set. While a bit strange, this makes sense for the setup we have, with
MailChimp grabbing feeds at the top of the hour and us, knowing that
deadline, potentially updating feed contents within the half-hour
before.

Alternatively, a good idea would be to add URL query parameter that
forces a cache refresh.